Move hook to AdminSidebar
